pyproject.toml
[tool.poetry]
name = "avymail"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
httpx = "^0.23.1"
aioboto3 = "^10.2.0"
requests = "^2.28.1"
click = "^8.1.3"
jinja2 = "^3.1.2"
pytz = "^2022.7"
tqdm = "^4.64.1"
cachetools = "^5.2.0"
email-validator = "^1.3.0"
prometheus-client = "^0.18.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
mypy = "^0.991"
types-requests = "^2.28.11.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
